Implemented Switching between Home team and Away Team
I've implemented being able to switch between home team and away when downs try to go over 4, but i've run into an issue trying to get the scoring to work. I thought it was as simple as switching a tag and I thought that worked, but I realized that it was just going off as i started the scene. Sending Dr. Samuel an email about it.